Add tests for root project name generation
- part of #15655
