+++ subprojects/core/src/main/java/org/gradle/initialization/DefaultSettings.java
@@ -112,7 +112,8 @@ public abstract class DefaultSettings extends AbstractPluginAware implements Set
 
     private static String getProjectName(File settingsDir) {
         if(contains(File.listRoots(), settingsDir)) {
-            return "generated " +  settingsDir.toPath().getRoot().toString().replace("[\\:/]", "") + " " +
+            String rootIndicator = settingsDir.toPath().getRoot().toString().replaceAll("[\\\\:\\/]*", "");
+            return "generated " + rootIndicator  + (rootIndicator.isEmpty() ? "" : " ") +
                 sha512().hashString(now().toString()).toString().substring(0, 6);
         }
         return settingsDir.getName();
