+++ subprojects/core/src/test/groovy/org/gradle/initialization/DefaultSettingsRootFileSystemTest.groovy
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2009 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.initialization
+
+import org.gradle.internal.os.OperatingSystem
+
+
+class DefaultSettingsRootFileSystemTest extends DefaultSettingsCommonTest {
+
+    def setup() {
+       createSettings('/')
+    }
+
+    def 'root project name generation in file system root'() {
+        expect:
+        settings.rootProject.name ==~ getRootProjectNamePattern()
+    }
+
+    private getRootProjectNamePattern() {
+        if(OperatingSystem.current().isWindows()) {
+            return 'generated [a-zA-Z] [a-z0-9]{6}'
+        }
+        return 'generated [a-z0-9]{6}'
+
+    }
+
+}
